@@ -65,7 +65,7 @@ const LoginPage = () => {
 			</div>
 			<div className='px-8 py-4 bg-gray-900 bg-opacity-50 flex justify-center'>
 				<p className='text-sm text-gray-400'>
-					Don't have an account?{" "}
+				Don&apos;t have an account?{" "}
 					<Link to='/signup' className='text-green-400 hover:underline'>
 						Sign up
 					</Link>
