@@ -1,112 +1,90 @@
 import { motion } from "framer-motion";
-import Input from "../components/Input"; // Assuming you have an Input component
-import { Loader, Lock, Mail, User } from "lucide-react"; // Icons
+import Input from "../components/Input";
+import { Loader, Lock, Mail, User } from "lucide-react";
 import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import PasswordStrengthMeter from "../components/PasswordStrengthMeter"; // Assuming you have a PasswordStrengthMeter component
-import { useAuthStore } from "../store/authStore"; // Zustand auth store
+import PasswordStrengthMeter from "../components/PasswordStrengthMeter";
+import { useAuthStore } from "../store/authStore";
 
 const SignUpPage = () => {
-  // Local state for form inputs
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
+	const [name, setName] = useState("");
+	const [email, setEmail] = useState("");
+	const [password, setPassword] = useState("");
+	const navigate = useNavigate();
 
-  // Navigation hook from react-router-dom
-  const navigate = useNavigate();
+	const { signup, error, isLoading } = useAuthStore();
 
-  // Zustand auth store, extracting signup method, error, and isLoading state
-  const { signup, error, isLoading } = useAuthStore();
+	const handleSignUp = async (e) => {
+		e.preventDefault();
 
-  // Form submission handler
-  const handleSignUp = async (e) => {
-    e.preventDefault();
+		try {
+			await signup(email, password, name);
+			navigate("/verify-email");
+		} catch (error) {
+			console.log(error);
+		}
+	};
+	return (
+		<motion.div
+			initial={{ opacity: 0, y: 20 }}
+			animate={{ opacity: 1, y: 0 }}
+			transition={{ duration: 0.5 }}
+			className='max-w-md w-full bg-gray-800 bg-opacity-50 backdrop-filter backdrop-blur-xl rounded-2xl shadow-xl 
+			overflow-hidden'
+		>
+			<div className='p-8'>
+				<h2 className='text-3xl font-bold mb-6 text-center bg-gradient-to-r from-green-400 to-emerald-500 text-transparent bg-clip-text'>
+					Create Account
+				</h2>
 
-    try {
-      // Call the signup function from the auth store
-      await signup(email, password, name);
-      // Navigate to the verify-email page on successful signup
-      navigate("/verify-email");
-    } catch (error) {
-      console.log("Signup Error: ", error);
-    }
-  };
+				<form onSubmit={handleSignUp}>
+					<Input
+						icon={User}
+						type='text'
+						placeholder='Full Name'
+						value={name}
+						onChange={(e) => setName(e.target.value)}
+					/>
+					<Input
+						icon={Mail}
+						type='email'
+						placeholder='Email Address'
+						value={email}
+						onChange={(e) => setEmail(e.target.value)}
+					/>
+					<Input
+						icon={Lock}
+						type='password'
+						placeholder='Password'
+						value={password}
+						onChange={(e) => setPassword(e.target.value)}
+					/>
+					{error && <p className='text-red-500 font-semibold mt-2'>{error}</p>}
+					<PasswordStrengthMeter password={password} />
 
-  return (
-    <motion.div
-      initial={{ opacity: 0, y: 20 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ duration: 0.5 }}
-      className="max-w-md w-full bg-gray-800 bg-opacity-50 backdrop-filter backdrop-blur-xl rounded-2xl shadow-xl overflow-hidden"
-    >
-      <div className="p-8">
-        <h2 className="text-3xl font-bold mb-6 text-center bg-gradient-to-r from-green-400 to-emerald-500 text-transparent bg-clip-text">
-          Create Account
-        </h2>
-
-        {/* Signup Form */}
-        <form onSubmit={handleSignUp}>
-          {/* Name Input */}
-          <Input
-            icon={User}
-            type="text"
-            placeholder="Full Name"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            required
-          />
-          
-          {/* Email Input */}
-          <Input
-            icon={Mail}
-            type="email"
-            placeholder="Email Address"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-          />
-          
-          {/* Password Input */}
-          <Input
-            icon={Lock}
-            type="password"
-            placeholder="Password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-          />
-
-          {/* Error Message */}
-          {error && <p className="text-red-500 font-semibold mt-2">{error}</p>}
-
-          {/* Password Strength Meter */}
-          <PasswordStrengthMeter password={password} />
-
-          {/* Submit Button */}
-          <motion.button
-            className="mt-5 w-full py-3 px-4 bg-gradient-to-r from-green-500 to-emerald-600 text-white font-bold rounded-lg shadow-lg hover:from-green-600 hover:to-emerald-700 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2 focus:ring-offset-gray-900 transition duration-200"
-            whileHover={{ scale: 1.02 }}
-            whileTap={{ scale: 0.98 }}
-            type="submit"
-            disabled={isLoading} // Disable button while loading
-          >
-            {/* Show loader while loading, else show "Sign Up" */}
-            {isLoading ? <Loader className="animate-spin mx-auto" size={24} /> : "Sign Up"}
-          </motion.button>
-        </form>
-      </div>
-
-      {/* Already have an account? */}
-      <div className="px-8 py-4 bg-gray-900 bg-opacity-50 flex justify-center">
-        <p className="text-sm text-gray-400">
-          Already have an account?{" "}
-          <Link to={"/login"} className="text-green-400 hover:underline">
-            Login
-          </Link>
-        </p>
-      </div>
-    </motion.div>
-  );
+					<motion.button
+						className='mt-5 w-full py-3 px-4 bg-gradient-to-r from-green-500 to-emerald-600 text-white 
+						font-bold rounded-lg shadow-lg hover:from-green-600
+						hover:to-emerald-700 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2
+						 focus:ring-offset-gray-900 transition duration-200'
+						whileHover={{ scale: 1.02 }}
+						whileTap={{ scale: 0.98 }}
+						type='submit'
+						disabled={isLoading}
+					>
+						{isLoading ? <Loader className=' animate-spin mx-auto' size={24} /> : "Sign Up"}
+					</motion.button>
+				</form>
+			</div>
+			<div className='px-8 py-4 bg-gray-900 bg-opacity-50 flex justify-center'>
+				<p className='text-sm text-gray-400'>
+					Already have an account?{" "}
+					<Link to={"/login"} className='text-green-400 hover:underline'>
+						Login
+					</Link>
+				</p>
+			</div>
+		</motion.div>
+	);
 };
-
 export default SignUpPage;
