@@ -19,11 +19,10 @@ export const useAuthStore = create((set) => ({
 			const response = await axios.post(`${API_URL}/signup`, { email, password, name });
 			set({ user: response.data.user, isAuthenticated: true, isLoading: false });
 		} catch (error) {
-			set({ error: error.response?.data?.message || "Error signing up", isLoading: false });
+			set({ error: error.response.data.message || "Error signing up", isLoading: false });
 			throw error;
 		}
 	},
-	
 	login: async (email, password) => {
 		set({ isLoading: true, error: null });
 		try {
@@ -50,7 +49,6 @@ export const useAuthStore = create((set) => ({
 			throw error;
 		}
 	},
-
 	verifyEmail: async (code) => {
 		set({ isLoading: true, error: null });
 		try {
@@ -58,11 +56,10 @@ export const useAuthStore = create((set) => ({
 			set({ user: response.data.user, isAuthenticated: true, isLoading: false });
 			return response.data;
 		} catch (error) {
-			set({ error: error.response?.data?.message || "Error verifying email", isLoading: false });
+			set({ error: error.response.data.message || "Error verifying email", isLoading: false });
 			throw error;
 		}
 	},
-
 	checkAuth: async () => {
 		set({ isCheckingAuth: true, error: null });
 		try {
@@ -72,7 +69,6 @@ export const useAuthStore = create((set) => ({
 			set({ error: null, isCheckingAuth: false, isAuthenticated: false });
 		}
 	},
-
 	forgotPassword: async (email) => {
 		set({ isLoading: true, error: null });
 		try {
@@ -81,12 +77,11 @@ export const useAuthStore = create((set) => ({
 		} catch (error) {
 			set({
 				isLoading: false,
-				error: error.response?.data?.message || "Error sending reset password email",
+				error: error.response.data.message || "Error sending reset password email",
 			});
 			throw error;
 		}
 	},
-
 	resetPassword: async (token, password) => {
 		set({ isLoading: true, error: null });
 		try {
@@ -95,7 +90,7 @@ export const useAuthStore = create((set) => ({
 		} catch (error) {
 			set({
 				isLoading: false,
-				error: error.response?.data?.message || "Error resetting password",
+				error: error.response.data.message || "Error resetting password",
 			});
 			throw error;
 		}
